Add files via upload
A portfolio introduction is pretty similar to a cover letter. In a cover letter, you might say "Please see attached resume" along with a few bullets on what to look for specifically. The same thing applies to a portfolio introduction—you could say "Please see my attached portfolio of examples of projects.